requirements.txt
backports.tarfile==1.2.0
bs4==0.0.2
chromadb==1.0.7
httptools==0.6.4
ipykernel==6.29.5
ipywidgets==8.1.6
jaraco.collections==5.1.0
pip-chill==1.0.3
plotly==6.0.1
python-dotenv==1.1.0
seaborn==0.13.2
sentence-transformers==4.1.0
streamlit-option-menu==0.4.0
tomli==2.0.1
umap-learn==0.5.7
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
